LayoutTests/fast/dom/call-a-constructor-as-a-function-expected.txt

This tests if TypeError is thrown or not when we call a constructor as a normal function.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ArrayBuffer() threw exception TypeError: ArrayBufferConstructor is not a function (evaluating 'ArrayBuffer()').
SKIP AudioContext is not implemented.
PASS FormData() threw exception TypeError: FormDataConstructor is not a function (evaluating 'FormData()').
PASS DataView() threw exception TypeError: DataViewConstructor is not a function (evaluating 'DataView()').
PASS EventSource() threw exception TypeError: EventSourceConstructor is not a function (evaluating 'EventSource()').
PASS FileReader() threw exception TypeError: FileReaderConstructor is not a function (evaluating 'FileReader()').
PASS Float32Array() threw exception TypeError: Float32ArrayConstructor is not a function (evaluating 'Float32Array()').
PASS Float64Array() threw exception TypeError: Float64ArrayConstructor is not a function (evaluating 'Float64Array()').
PASS Audio() threw exception TypeError: AudioConstructor is not a function (evaluating 'Audio()').
PASS Image() threw exception TypeError: ImageConstructor is not a function (evaluating 'Image()').
PASS Option() threw exception TypeError: OptionConstructor is not a function (evaluating 'Option()').
PASS Int16Array() threw exception TypeError: Int16ArrayConstructor is not a function (evaluating 'Int16Array()').
PASS Int32Array() threw exception TypeError: Int32ArrayConstructor is not a function (evaluating 'Int32Array()').
PASS Int8Array() threw exception TypeError: Int8ArrayConstructor is not a function (evaluating 'Int8Array()').
PASS MessageChannel() threw exception TypeError: MessageChannelConstructor is not a function (evaluating 'MessageChannel()').
PASS SharedWorker() threw exception TypeError: SharedWorkerConstructor is not a function (evaluating 'SharedWorker()').
PASS Uint16Array() threw exception TypeError: Uint16ArrayConstructor is not a function (evaluating 'Uint16Array()').
PASS Uint32Array() threw exception TypeError: Uint32ArrayConstructor is not a function (evaluating 'Uint32Array()').
PASS Uint8Array() threw exception TypeError: Uint8ArrayConstructor is not a function (evaluating 'Uint8Array()').
PASS WebKitCSSMatrix() threw exception TypeError: WebKitCSSMatrixConstructor is not a function (evaluating 'WebKitCSSMatrix()').
PASS WebKitPoint() threw exception TypeError: WebKitPointConstructor is not a function (evaluating 'WebKitPoint()').
PASS WebSocket() threw exception TypeError: WebSocketConstructor is not a function (evaluating 'WebSocket()').
PASS Worker() threw exception TypeError: WorkerConstructor is not a function (evaluating 'Worker()').
PASS XMLHttpRequest() threw exception TypeError: XMLHttpRequestConstructor is not a function (evaluating 'XMLHttpRequest()').
PASS XSLTProcessor() threw exception TypeError: XSLTProcessorConstructor is not a function (evaluating 'XSLTProcessor()').
PASS successfullyParsed is true

TEST COMPLETE
